% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess.R
\name{running_ess}
\alias{running_ess}
\title{Running effective sample size}
\usage{
running_ess(w, f, x)
}
\arguments{
\item{w}{A numeric vector of non-negative weights.}

\item{f}{A function used in computing \code{f}-specific ESS.}

\item{x}{A numeric vector of samples used to generate \code{w}. Used for computing \code{f(x)}.}
}
\value{
An effective sample size estimate.
}
\description{
Computes and returns the running estimate of effective sample size (ESS) of 
importance sampling estimator.
}
