% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{running_mean}
\alias{running_mean}
\title{Compute running mean of a vector}
\usage{
running_mean(x, na.rm)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{If \code{TRUE}, \code{NA} values in \code{x} are
omitted from the computation. Default is \code{FALSE}.}
}
\value{
A vector containing the recursive mean estimates.
}
\description{
Computes running mean of a vector, returning the values from each step.
}

