% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\name{running_weighted_var}
\alias{running_weighted_var}
\title{Compute running weighted variance of a vector}
\usage{
running_weighted_var(x, w, na.rm = FALSE, method = c("unbiased", "moment"))
}
\arguments{
\item{x}{A numeric vector.}

\item{w}{A numeric vector of non-negative weights. Will be automatically normalized to sum to one.}

\item{na.rm}{If \code{TRUE}, \code{NA} values in \code{x} (and corresponding weights in \code{w}) are
omitted from the computation. Default is \code{FALSE}.}

\item{method}{Estimator type, either \code{"unbiased"} (default) or \code{"moment"}. In unbiased case the 
sum of squares is divided by \code{length(x) - 1}, whereas in the latter case the division is by \code{length(x)}.}
}
\value{
A vector containing the recursive weighted variance estimates.
}
\description{
Computes running weighted variance of a vector, returning the values from each step.
}

