% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{weighted_mean}
\alias{weighted_mean}
\title{Compute a weighted mean of a vector}
\usage{
weighted_mean(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{w}{A numeric vector of non-negative weights. Will be automatically normalized to sum to one.}

\item{na.rm}{If \code{TRUE}, \code{NA} values in \code{x} (and corresponding weights in \code{w}) are
omitted from the computation. Default is \code{FALSE}.}
}
\value{
A weighted mean.
}
\description{
Computes a weighted mean of a vector.
}

