% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses.R
\name{weighted_se}
\alias{weighted_se}
\title{Weighted standard error}
\usage{
weighted_se(x, w, na.rm)
}
\arguments{
\item{x}{A numeric vector or matrix.}

\item{w}{A numeric vector of non-negative weights. Will be automatically normalised to sum to one.}

\item{na.rm}{If \code{TRUE}, \code{NA} values in \code{x} (and corresponding weights in \code{w}) are
omitted from the computation. Default is \code{FALSE}.}
}
\value{
A weighted variance.
}
\description{
Computes a weighted standard error of a vector or matrix.
}
\note{
Compared to some other R functions, here the weights are regarded as probability weights,
not frequency weights.
}
