% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{running_mean}
\alias{running_mean}
\title{Running mean}
\usage{
running_mean(x, na.rm)
}
\arguments{
\item{x}{A numeric vector, matrix or three dimensional array.
For matrix, the mean is computed for each column, and 
for array the sweep is done over the third dimension.}

\item{na.rm}{If \code{TRUE}, \code{NA} values in \code{x} are
omitted from the computation. Default is \code{FALSE}.}
}
\value{
A vector containing the recursive mean estimates.
}
\description{
Computes running mean of a vector or matrix, returning the values from each step.
}

