% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\name{running_var}
\alias{running_var}
\title{Running variance of a vector}
\usage{
running_var(x, method = c("unbiased", "moment"), na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{method}{Estimator type, either \code{"unbiased"} (default) or \code{"moment"}.}

\item{na.rm}{If \code{TRUE}, \code{NA} values in \code{x} are
omitted from the computation. Default is \code{FALSE}.}
}
\value{
A vector containing the recursive variance estimates.
}
\description{
Computes running variance of a vector, returning the values from each step.
}

