% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\name{running_var}
\alias{running_var}
\title{Running variance of a vector}
\usage{
running_var(x, method = c("moment", "unbiased"), na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector or object that can be coerced to such.}

\item{method}{Estimator type, either \code{"moment"} (default) or 
\code{"unbiased"}, which is unbiased only in case of frequency weights.}

\item{na.rm}{If \code{TRUE}, \code{NA} values in \code{x} are
omitted from the computation. Default is \code{FALSE}.}
}
\value{
A vector containing the recursive variance estimates.
}
\description{
Computes running variance of a vector, returning the values from each step.
}
