% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{running_weighted_mean}
\alias{running_weighted_mean}
\title{Compute running weighted mean of a vector}
\usage{
running_weighted_mean(x, na.rm)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{If \code{TRUE}, \code{NA} values in \code{x} (and corresponding weights in \code{w}) are
omitted from the computation. Default is \code{FALSE}.}
}
\value{
A vector containing the recursive weighted mean estimates.
}
\description{
Computes running weighted mean of a vector, returning the values from each step.
}

