% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{running_weighted_mean}
\alias{running_weighted_mean}
\title{Running weighted mean}
\usage{
running_weighted_mean(x, w, na.rm)
}
\arguments{
\item{x}{A numeric vector, matrix or three dimensional array.
For matrix, the mean is computed for each column, and 
for array the sweep is done over the third dimension.}

\item{w}{A numeric vector of non-negative weights. Will be automatically normalised to sum to one.}

\item{na.rm}{If \code{TRUE}, \code{NA} values in \code{x} (and corresponding weights in \code{w}) are
omitted from the computation. Default is \code{FALSE}.}
}
\value{
A vector containing the recursive weighted mean estimates.
}
\description{
Computes running weighted mean of a vector or matrix, returning the values from each step.
}

