% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\name{running_weighted_var}
\alias{running_weighted_var}
\title{Running weighted variance of a vector}
\usage{
running_weighted_var(x, w, method = c("moment", "unbiased"), na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector or object that can be coerced to such.}

\item{w}{A numeric vector of non-negative weights. Will be automatically 
normalised to sum to one.}

\item{method}{Estimator type, either \code{"moment"} (default) or 
\code{"unbiased"}, which is unbiased only in case of frequency weights.}

\item{na.rm}{If \code{TRUE}, \code{NA} values in \code{x} (and corresponding weights in \code{w}) are
omitted from the computation. Default is \code{FALSE}.}
}
\value{
A vector containing the recursive weighted variance estimates.
}
\description{
Computes running weighted variance of a vector, returning the values from each step.
}
