% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\name{weighted_var}
\alias{weighted_var}
\title{Compute a weighted variance of a vector}
\usage{
weighted_var(x, w, method, na.rm)
}
\arguments{
\item{x}{A numeric vector, matrix or array.}

\item{w}{A numeric vector of non-negative weights. Will be automatically normalized to sum to one.}

\item{method}{Estimator type, either \code{"unbiased"} (default) or \code{"moment"}.}

\item{na.rm}{If \code{TRUE}, \code{NA} values in \code{x} (and corresponding weights in \code{w}) are
omitted from the computation. Default is \code{FALSE}.}
}
\value{
A weighted variance.
}
\description{
Computes a weighted variance of a vector.
}

